requirements.txt
google-adk==1.14.0
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
python-dotenv==1.1.0
PyPDF2==3.0.1
python-docx==1.1.2
Pillow==11.2.1
google-cloud-aiplatform==1.90.0
